pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lsl_xdf_now"
version = "0.0.1"
description = "Repository for handling LSL streams and XDF data."
license = {text = "Proprietary"}
authors = [
    { name = "Henesis Team", email = "m.barsotti@henesis.eu" }
]
requires-python = ">=3.12"
dependencies = [
    "pyyaml",
    "mne==1.4.2",
    "numpy>=1.20.3",
    "scipy",
    "plotly",
    "flask",
    "matplotlib",
    "seaborn",
    "einops>=0.3.0",
    "scikit-learn",
    "pylsl==1.13.6",
    "fire",
    "numba",
    # "PyQt5",
    # "pyqtgraph",
    # "sounddevice"
]
classifiers = [
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.optional-dependencies.dev]
lint = ["pylint"]

[project.scripts]
start-viz-gui = "scripts.start_gui:main"

[tool.setuptools.packages.find]
where = ["lsl_xdf_now"]
exclude = ["tests", "scripts"]
